Removed option for immediate commits in dsShaderVariableGroup.
Less efficient for large arrays of data, but no longer requires a command
buffer for dsShaderVariableGroup_setElementData(). This guarantees
materials can be set up on separate threads when a resource context is
available but a command buffer is not. This also removes the situations
where dsShaderVariableGroup_setElementData() could allocate memory.