Fixed scaling and padding of text.
Calculation of bounds is now accurate when the glyph needs to be scaled
when placing into the texture. Scaling is no longer inverted.
Added an extra quality level between medium and high (now very high).
Added an explicit None cube face, which is an alias to PosX. This is
clearer when not using cube maps.