Support display vs. pixel coordinates.
Can get the window pixel size separate from the window size in display
coordinates. Attempting to work in pure pixel coordinates on Mac is a
losing battle, so it's up to the clients ot do the conversion when
necessary. The display bounds are now available so sensible window sizes
can be created, rather than attempting to use the resolution which could
have incorrect positions and sizes.