Added more triangulation tests.
Added sawtooth pointing to the left, polygon with a hole, and not closing
the path. Triangulation now handles corners where there could be an left
or right edge both above and below the corner. The detection of neighboring
edges to split is now purely geometric based on the angles with respect to
the added edge. (required for more than two added edges)