Implemented more correct bisecting edge detection.
Detecting the top and bottom edges was correct, but searching for a closer
point could either detect a point that passes through the polygon or miss
a point it should have caught. Now search for the closest point on the
left or right that doesn't intersect any other edges. Also use a BVH to
search for intersections to improve the time complexity.

Added unit tests for the problem cases.