Fixes for SimplePolygon.
Simplified the computation of the winding order, and make sure the proper
vertices are used.

Now use a list for connected edges rather than a hard-coded array. There
were some cases where there weren't enough slots to properly connect the
vertices for the edges.

Added stress tests for simple polygons to give an idea of performance. The
starburst test also helped discover errors in the algorithm.