Added test for polygon that has a hole surrounded by another.
dsSimplePolgon now better handles touching vertices. When connecting edges,
makes sure that the connection is internal to the polygon.

Improved how loops are connected with dsSimpleHolePolygon. It now uses set
theory to try and connect each group of loops, and doesn't add redundant
connections when vertices touch.

When traversing edges, always try to take the edge with the least angle.
If it's already traversed, the traversal ends. This should avoid loops that
touch on the inside, which would attempt to create a degenerate connecting
edge.