Added offset to text style and fixed line wrapping.
Can now apply a vertical offset to text. (e.g. for superscript) Don't
support a horizontal offset since it interferes with line wrapping, and
there's no current need for it.

Fixed logic for determining breaks for line wrapping. Added a limit line
to TestText for strings that contain a max width to more accurately
determine if line wrapping works correctly.