Added code to handle vector image text.
It's now a requirement to provide a command buffer when creating a
dsVectorImage. Added dsVectorImage_updateText() to refresh the text and
re-add the ranges to the render buffers.

Removed command for path text. To be implemented correctly it would need to
cache the transforms for each glyph to be re-applied each time the text is
updated, ensuring proper management of glyph caching and texture
coordinates. This would also need to be passed through to the shader in the
vertex buffer for tessellation shaders, requiring a separate shader for
path and non-path text. (or adding extra data for non-path text) Given that
text along paths is somewhat niche, it's better to drop support for now.