Polygon edge intersection now uses proper ranges.
When intersecting two edges, make sure that the T values are in the proper
ranges for both edges rather than just the first one. The first edge is
exclusive of the end points, while the second edge is inclusive. This means
that the first edge is allowed to touch other edges, but not vice-versa.