More improvements to text management.
Updated SheenBidi. Now determine directionality directly rather than based
on the first ordered codepoint. (except for uniform scripts) This is based
off the run level, which will be odd for right to left runs. This is
determined by the Unicode BiDi algorithm standard, but was clarified in the
SheenBidi documentation as of the latest version.

Script transitions are now always detected based on original character
order. Right to left sections are reversed as a post-process to ensure
proper ordering. This now properly detects inherited scripts for script
boundaries. Added test for inherited text between script boundaries.

Newlines are now properly detected with mixed right to left and left to
right text. This is done by first detecting explicit newlines from ranges
before following character order processing for wrapping. This is because
right to left ranges will be reversed, but only up to line boundaries.

Fixed cases with uniform script, and added test cases for both left to
right and right to left uniform scripts. When detecting the first
codepoint, ignore common text.