Many improvements and fixes for profiling.
Enable profiling for TestVectorImage and TestText. Fix various issues based
on viewing the results.

Added error checking for profiling push/pop scopes. There is no useful
error checking in easy_profiler, and it wouldn't help you if profiling
isn't enabled. When debugging is enabled, it will verify the consistency
of the scopes, ensuring pushes and pops are properly matched. In the case
of functions, it also verifies that the push and corresponding pop are in
the same function. When an error occurs, it will print out information
about the currently pushed scopes, break in the debugger, and abort. This
checking is enabled even when profiling functions haven't been set, so
cases like unit tests will still perform the checks in debug mode.