dsCommandBuffer now holds bound objects.
Core renderer will now do sanity checks about state, such as making sure
a render pass or shader is active, or that render passes are properly
managed. This also replaces the hacky thread-local variable previously used
to verify render pass state, which could also get stuck in a bad state.

Much of the error checking in the OpenGL implementation has been moved to
asserts. The core logic handles the error messages for the API user, while
the internal state is kept as a shadow state to manage cases such as
submitting one command buffer to another. (which bypasses the main command
buffer state)

Framebuffer now holds a name, which will be used for profiling.