Enforce command buffer operations are done inside a frame.
This must be inside a begin/end frame block or in a secondary command
buffer. This will be important for Vulkan as well as to guarantee
consistency when hooking up GPU times.

TestText and TestVectorDraw now use a secondary command buffer for setup.
This avoids having to create a dummy frame and also is a practical test for
secondary command buffers.