Added checks for what supported renderers and devices.
Added isSupported() functions for renderers to check if they are supported.
RenderBootstrap can now determine the default renderer ahead of time.

Added queryDevices() functions for renderers. Gets the list of devices that
are available to use. The UUID of the device may be set on the
dsRenderOptions structure to choose which one to use.