WIP for Vulkan intialization.
Create Vulkan instance, device, and queue. Register debug callback when
requrested to print errors, complete with callsite information.

Embed Vulkan headers rather than trying to find them on the system. Since
the library is dynamically loaded, and the existing API will remain stable,
this removes the requirement to install the SDK on the building system. In
some cases (such as Windows) the library may be present but not the headers
by default.