Fixes for building on Android.
Ignore DEEPSEA_OUTPUT_DIR on Android. Android Studio will supply the
library output directory, and changing it will break other parts of the
build.

Added ds_add_executable_target() CMake function. On Android this will build
a shared library rather than an executable, ignoring the WIN32 and
MACOSX_BUNDLE options.

Text should now build when statically linking FreeType and HarfBuzz.
Explicitly link to zlib, specify proper linking order, and add a .cpp file
to force C++ linkage for the library.

Added ds_build_assets_dir() CMake function to determine the directory to
build assets. For most platforms, this tries to have the same directory as
$<TARGET_FILE_DIR:${target}> would have done before. On Android this
computes the intermediate assets directory so it will be embedded in the
APK. While less robust in finding the directory than the previous method,
this allows the executable target to depend on the copy target rather than
the other way around. This is a requirement to bundle the assets on
Android, but also is better for other platforms since it doesn't require an
all build to guarantee that the assets get copied.