Fixes for Android.
Output assets to project directory rather than intermediate directory. This
broke with the recent Android Studio 3.2 update, and while it would be nice
to place them in the build directory, putting it as part of the project
should be more robust.

Fixed with flag in ApplicationSDL.

Manually check for resizes based on surface size. This fixes changing
orientation on Android often returning the old size.