Added command buffer management for Vulkan.
Create a pool of command buffers as a round-robin. As we come up to a
command buffer, use a fence to ensure it's unused before we continue.
Have a separate command buffer for resource operations and render commands,
allowing for resources discovered to be used are processed first. The
submit index is tracked can wait until a submit index is has been
processed, allowing synchronization on the host.

Added some initial resource processing and safe deletion code for buffers.