Take DPI into account for testers.
Added dsApplication_adjustWindowSize() to adjust DPI for windows based on
the operating system. Adjust the size of the windows for the existing
testers, and adjust the size of text in TestText so it scales properly.

This makes testers usable on very high DPI screens, such as a 4k laptop
screen. For TestText, the text should now be readable on a phone screen.