Implemented texture creation for Vulkan.
Can now choose to resolve a single surface or all surfaces for offscreens.
This determines how large to make the surface to render to on Vulkan.
Added helper functions for the number of surfaces on a texture and the
index of a surface. This is used when needing to keep an array of
individual surfaces inside of a texture.