Use single image for initial texture data when possible.
Check to see if the device supports the parameters for a linear image, and
if so, use a single image to hold the data. This should improve initial
upload performance for devices that support it, while falling back to the
image per surface approach for devices that don't.

Implemented deletion of Vulkan textures.