Use semaphore for command buffers with render surfaces.
Keep track of render surfaces used with command buffers to wait on for
submit. Look for duplicates for both render surfaces and readback
offscreens so they don't get double processed, and also since the number is
expected to be small.

Use lazy allocation when possible for extra render surface images.