Improvements to Vulkan resource processing.
Now uses dsLifetime objects to ensure that the object isn't deleted before
it's processed. Added processing for renderbuffers for the required layout
transition, and added calls to ensure that resources are processed before
they are used for various operations.

Added functions to force processing of resources for textures and buffers
immediately rather than on first use. This is a hint for APIs that support
it (currently Vulkan) and isn't guaranteed to do anything.