Vulkan render passes now handle setting AA samples.
When the default samples are used, it will now re-create the underlying
render pass. The lifetime is now properly managed, and will manage the
framebuffers associated with it, allowing for old unused framebuffers to
be destroyed.

Number of attachments is now deterministic: 2x the original number to
account for multisample resolve. This allows for render passes to be used
regardless of if the default surface has anti-aliasing or not. Also set up
explicit resolving of the main framebuffer. Fixed handling of stereoscopic
render surfaces as well.