Fixes for Vulkan render surfaces and render passes.
Choose the color surface format ahead of time. It will choose a BGRA format
and force alpha for better compatibility. While this isn't 100% guaranteed
to have a compatible format, it's the only way to have the format in the
renderer match the Vulkan format on the surface so render passes and
framebuffers are compatible. BGRA should be the most commonly supported
format, but you can now set a forced format in dsRendererOptions for
specialized platforms that need a specific format.

Render passes no longer have resolve surfaces for all attachments. It now
only creates them for attachments with the resolve flag set, and it's
validated that framebuffers have a resolvable surface in this case. The
previous fallback behavior would fail due to sample mismatches.