Optimize offscreens that are only subpass inputs.
Keep the offscreen as the optimal layout for color or depth/stencil
attachments rather than shader read optimal. This prevents the need for a
barrier before and after the render pass for the layout transition, and
will only have the implicit layout transition within the subpass it's used
as an attachment for.