Improved store ops for render pass attachments.
Fixed subpass indices for TestRenderSubpass shader. They were changed at
one point thinking it was the original attachment index when trying to get
it to work in the first place, but it's the index within the input
attachment array for that subpass in isolation. Fixes crashes on drivers
that actually use it.