Reverted reduced calls to dsRenderSurface_update().
On Android there's a delay between the window change and underlying surface
change, which causes it to get an incorrect size. The original motivation
for this change was due to long delays observed on Linux, but that seems to
be for *any* surface operation caused by the compositor on KDE rather than
this call specifically.