Fixes for TestRenderSubpass on Mali.
More properly handle lazy allocated flag. No longer specify blit flag for
depth buffers so they may actually be lazily allocated.

Application library will now register a wrapper log function to show a
dialog and exit on a fatal error. This is only used if the log function
hasn't been overridden. This allows for fatal errors to be communicated to
the user, such as the device lost error that sometimes shows up on Mali
when resizing the window. (I've done as much synchronization as I can to
try to prevent this, but at this point given that there's no validation
errors it seems like a driver issue)