Fixes for testers on Mali.
Use new varying syntax for shaders to avoid input and output structs, which
the driver doesn't like. Vector draw shaders are still to be done.

Allow for delay in creation of window surface. This allows for the window
to be set fullscreen and avoid a resize on the window, working around a pop
on startup and potential driver errors with surface resizing.

Don't free resources still associated with a command buffer when waiting
for idle. Only resources that will no longer be in use are destroyed.

Workaround for the colorspace enum not working for OpenGL ES surfaces on
Mali despite having the colorspace extension.