Fixes for Mac.
Ignore deprecation warnings. In the case of OpenGL functions, I won't be
using the deprecated functions, and they will be NULL due to weak linking
during loading. In the case of GLPlatformCocoa.m, it says to use
NSOpenGLView which simply isn't going to happen. I will likely have a Metal
implementationbefore this becomes an actual problem anyway.

Fix for broken window resizing on Mac.