Improved handling of paths.
Added function to query if a path is absolute.

Added "External" resource stream type, which is for files external to the
project. The path is untouched when opening.

Added dsResourceStream_getPath(). Gets the path to the resource that would
be used to open the file. The dsFaceGroup_loadFaceResource() function now
uses this to get the path rather than combining with the directory name
directly. This also checks if the input path is absolute, returning an
error if an absolute path is used with anything but external resources.