Renamed dsVolatileMterialValues to dsSharedMaterialValues.
This name is more descriptive for the usage of the structure. While the
values are able to spontaneously change, doing so can be expensive in some
situations. (e.g. creating a new descriptor set for all but offset changes
on Vulkan) Most uses will be for shared systems like transforms or lighting
values, so "shared" is more appropriate.