Improved consistency for naming of texture buffers.
ImageBuffer -> TextureBuffer: This is in line with textures, which are
immutable.
MutableImageBuffer -> ImageBuffer: This is in line with images, which are
mutable.

While in general the distinction between texture and image is a bit
strange, it's already established and it brings texture buffers more in
line with existing texture/image terminology.