Optimized setting of push constant data.
Cached the push constant mapping to material elements and metadata so it
doesn't need to be looked up each time a shader is bound with a material.
Also fixed management of arrays, both getting the correct array length and
handling the case where the stride differs from the element size.