Fixes for Vulkan submit management.
Make sure access to finishedSubmitCount is always protected by the mutex
since it can change across threads whenever the sumbit is waited on.

Make sure that the value of finishedSubmitCount is actually less than the
value it's being assigned to during flush. Otherwise a call to
dsVkRenderer_waitForSubmit() could have been made (e.g. when waiting for
the swap chain image) with a much higher value that would have been
overwritten, resulting in unnecessary waits or thinking when something
isn't finished when it is.

Calling dsRenderer_waitUntilIdle() will now update the finished submit.