Re-worked command buffers for Metal.
Separate command buffer into a "software" command buffer, which has an
internal memory buffer for commands, and "hardware" command buffer, which
goes straight to a Metal command buffer. A virtual function table is used
to go from the base command buffer to the specific implementation. Since
multi-submit and multi-frame command buffers can't be implemented by using
Metal command buffers directly, and support the full set of commands (as
opposed to the subset for secondary commands), a more proper separation
between the two command buffer types is warranted.

Command buffer pool now determines if hardware or software command buffers
need to be created based on the usage. If secondary, mult-submit, or
multi-frame a software command buffer is required.