Fixes for memory alignment when hashing.
Memory like string literals aren't guaranteed to be aligned, so performing
murmur3 hashing can crash due to unaligned access when not supported. (e.g.
on 32-bit ARM) Curiously this only appeared when optimization was enabled,
leading to the incorrect assumption that it was an optimization bug rather
than the pointer being unaligned.

Check for the alignment before hashing arbitrary bytes, using memcpy rather
than an assign. On some platforms these are identical performance, on
others it has a minor (~1-2%) performance hit.