Enable Vulkan push groups when profiling.
Check for DS_PROFILING_ENABLED rather than DS_GPU_PROFILING_ENABLED for
push groups. This allows external tools to be used, even if built-in GPU
profiling with timers isn't enabled. In this case, load the debug
extension, but don't enable the validation layers. This should avoid
performance hits. In this case, only debug labels are used. (avoiding
crashes in vkCreateDebugUtilsMessengerEXT() when using RenderDoc)