Finished hooking up Metal shaders.
Updated tools version and ModularShaderLanguage submodule for the required
extra Metal support.

Strip symbols and optimize TestVectorDraw shaders to verify that
functionality works. Rename STRIP_SYMBOLS argument to STRIP to be less
wordy and more consistent with other CMake functions.

Update script now recursively updates submodules. While this may be slower,
it should be more robust. Hopefully it will no longer need to be run twice
when the ModularShaderLanguage submodule is updated with newer versions
of glslang and SPIRV-Cross.