Fix for showing notifications on Android.
SDL activity paused the application on focus loss, which happens when
swiping down  notification bar. When the app is actually paused, the
surface is invalidated and needs to be re-created, but attempting to do
this on focus loss breaks. (Vulkan fails to create the surface since one
previously existed on that window)

Apart from this, it's better to have animations continue since the window
is still visible.