Improve Vulkan barriers.
Fix for incorrect barriers when copying between textures.

Buffers that can't be mapped now have barriers after copy. This avoids the
need to add a barrier each usage for the common case where only the command
buffer is used to copy data.

Fixed dsTextRenderBuffer to set the buffer as GPU only to indicate it can't
be mapped to. Per-frame barriers are no longer required for text rendering.