Simplified scene graph node management.
Instead of having a child ID for sibblings of the same node, don't allow
a node to be its own sibbling. This use case requires a unique intermediate
node, which would generally be the case regardless since each would need
its own transform to not draw on top of each-other.