Added dsSceneResources.
Holds resources that can be used within a scene indexable by name. The
lifetime is reference counted to ensure that nodes that use the resources
cannot be freed before the resources themselves. If managed properly this
should be the equivalent of manually freed resources, but if managed
improperly it will cause extra resources to be alive rather than a crash.