Fix for handling of gl_FragCoord with offscreens.
In OpenGL, don't flip gl_FragCoord when drawing to an offscreen since the
clip position (for gl_Position) is already inverted. As a result,
layout(origin_upper_left) cannot be used to avoid arithmatic in the shader
since the decision for whether or not to invert must be made at runtime.