Fixes for Metal.
Don't set bytesPerRow when writing PVR texture data. This is a nice
undocumented requirement that you don't know until you try to load a PVR
texture and it crashes.

Check for support of drawing with a base instance, which is also required
to draw with a vertex offset. When not supported, the vertex offset is set
via an offset to the vertex buffer. The strange alignment rules for vertex
buffers shouldn't be an issue here since only iOS doesn't support base
instances/vertices, which only requires 4 byte alignment.

Check for support of indirect draw and dispatch. Only compile in border
color support when targeting macOS 10.12+.