Optimizations for Metal.
Added dsMTLTempBuffer to hold temporary buffer data for push constants and
copy operations on the command buffer. Each buffer is 512 KB, and
allocations for copies larger than 256 KB go through the old behavior of
creating a temporary buffer or texture. Temp buffers are maintained in a
pool, and getting new buffer data first searches for an existing buffer
that's finished executing on the GPU.