Implemented creation of views.
Implemented creation and management of surfaces and framebuffers for views.
Surfaces are managed by name. They can have a pre-created surface or
information to dynamically create surfaces. Framebuffers reference surfaces
by name, and they are connected to the scene pipeline by name as well. When
the view is resized or default surface samples change, surfaces and
framebuffers are re-created as necessary.