Fixes for building for Android on Windows.
Fix for path for EGL library. Even if it built on Linux, it was likely
trying to link to the EGL library on the system, and probably just
coincidentally worked.

Always use UNIX-style paths for vector resource flatbuffers. This allows
cross-compiled files built on Windows to be used on targets such as
Android while the standard file IO functions on Windows can still handle
the UNIX-style paths when running natively.