Fixes for Vulkan barriers during resource processing.
When adding the "post" buffer for copies, check if dynamic mapping is
needed rather than if it's static. Also add a barrier before the copy for
the device buffer if it's not the initial copy.

Renamed dsVkGfxBufferData_canMap() to dsVkGfxBufferData_canMapMainBuffer()
to make it clearer that it's for the main buffer used for rendering rather
than being able to map at all. Also renamed parameters and local variables.

Make sure read as well as write access is protected for the "source" of
barriers. Also removed the "host" parameter for image barriers since
textures now use buffers to upload the data.