Hooked up clip and cull distances.
Checks that the number of clip and cull distances are valid when creating
a shader. OpenGL also needs to enable/disable clip distances based on the
render state.

Fixed checks of compute shader working group size. Added separate limits
for local working group size declared in the shader, which is checked when
loading the shader.