Various Vulkan fixes and improvements.
Fixed barriers for explicit clearing of surfaces, both for missing and
incorrect access and pipeline flags.

When generating mipmaps, undefined layout is used for the mip levels that
are being written to. This can potentially avoid some unnecessary work by
the GPU.

Buffer <-> image copies now enforce that the buffer dimensions are a
multiple of the block size. Unlike the texture dimensions, where the image
dimensions isn't necessarily a multiple of the block size, the buffer width
and height are for the data sizes.