Added support for more optimal render surface rotations.
Allow for client-side rotations, which are faster than compositor rotations
for cases like rotating a phone or tablet screen. When requested (through
clientRotations parameter for dsRenderSurface_create() or
dsWindowFlags_ClientRotations), the rotate field is set on dsRenderSurface.
The matrix for the rotation can be queried with
dsRenderSurface_makeRotationMatrix() and applied to the projection matrix
to perform the rotation. Note that viewport ranges and blit regions may
need to be adjusted as well. (this can be observed when rotating
TestRenderSubpass, since the regions don't rotate with the image as
expected)