Expose both stage and access flags for barriers.
Allows for more flexibility, especially for subpass inputs. This is the
first step towards unifying the memory model between OpenGL and Vulkan,
which both makes barriers more consistent between the two for dependencies
with shaders and will simplify the Vulkan implementation.

Added macro to declare bitmask operators for enums. This allows bitmask
enums to be used naturally in C++ without having to do casts. These macros
do nothing in C.