Improvements for explicit resolves.
Will now work properly for offscreens that are only used as subpass inputs
on Vulkan. Now use a separate flag rather than depending on
dsTextureUsage_OffscreenContinue, which is more explicit and obvious as a
requirement.

Don't allocate a separate array of bools to track resolve for Vulkan,
instead just re-use the attachment info array from the base render pass.