Change vertex format divisor to instanced bool.
Vulkan doesn't support a step rate other than 1 without an extension, which
is rather new and doesn't have much support on mobile devices. The previous
attempt to emulate the behavior was incorrect, since it's supposed to be
the number of instances to pass before incrementing the value rather than
the number of vertices to skip.