Changes for render pass multisample resolving.
No longer support an exlicit resolve at the end of the render pass. This
was done to support resolving depth/stencil surfaces on Vulkan, but is not
actually supported with vkCmdResolveImage. Instead added support for
resolving within a subpass like color surfaces, using the
VK_KHR_depth_stencil_resolve when available. Added
hasDepthStencilMultisampleResolve flag to dsRenderer to determine when
multisample resolving is supported.

Improved detection of what attachments need to be preserved for Vulkan
subpasses. This looks for individual attachment usage both before and after
the subpass by dependency order to determine when an attachment is written
to earlier and read from later.

Removed explicit resolve usage flag for textures. Also made it so copy to
flag is required to explicitly clear offscreens, which is more in line with
the requirements for Vulkan.

This currently works in Vulkan and OpenGL. Metal won't even compile yet,
and will be fixed in the next commit.