Added arbitrary data storage for item lists on nodes.
Each tree node now holds a dsSceneNodeItemData instance, which stores a
pointer for each item list the node is set on. Cull results now use this to
store the result value, allowing for separate values if the same node is
present under multiple transforms.

Removed dsSceneCullManager since this is superseded by the new way to store
results.