Clear the last used descriptor set periodically.
When the used resources are cleared in Vulkan command buffers, also clear
the last used descriptor set. This generally means at the end of a frame or
after submitting a command buffer to another.

This protects against situations where the comparison for dirty descriptor
sets turns out to be incorrect. For example, if a buffer or texture is
freed and re-allocated with the same memory address, but the contents,
including the Vulkan buffer, are different. This could be observed in
TestScene on Android when rotating the screen from vertical to horizontal
immediately after startup.