Made GPU profiling more robust.
Copy the name rather than storing a pointer. The objects submitting names
used to keep a pointer to the original name assuming it was a literal, but
it has since changed to copy the name to a buffer allocated with the
object. This means that if an object like a render pass is destroyed during
runtime, the name would be invalid while still used by the profile context.

Temporary hash table is now cleared after use. There were situations where
the nodes could change and cause major problems with the internal
structure.

Internal name buffers for nodes are now multiples of 8 to reduce padding
for 64-bit systems. Also finally added a reference to the Scene library in
the readme.