Vulkan optimizations.
Optimized dsGfxBuffer_copyData() implementation to use a temporary buffer
on Qualcomm, since the drivers take 1/3 ms to copy 200 B (!). This is also
used if multiple calls to vkCmdUpdateBuffer() would be required.

Optimized setting of transforms to not do operations on the GPU memory.
This greatly reduces the amount of time spent for some drivers.

Can now initialize EasyProfiler without beginning capture immediately. This
prevents memory buildup if startup doesn't need to be profiled.