Optimized binding of descriptors.
Rather than keeping a separate list of DeepSea objects for comparisons,
followed by creating the Vulkan bindings after, create the Vulkan bindings
ahead of time. This halves the amount of time taken when the bindings
change.

All texture and buffer elements are now required to be non-NULL. This
allows bindings to be pre-computed for Vulkan, and would be dependent on
a specific shader not using that binding to not potentially crash. Added a
workaround to VectorImage to ensure that all textures are set.