Fix for Android failures in TestScene.
When the render surface is invalidated, the pointer can be the same after
destroying and re-creating it. This can cause the screen to become black
when turning off the screen and turning it back on due to not updating the
framebuffer.

Remove the surface pointer check when setting a surface on a View. This
means that you need to be more careful when setting the surface, but it's
expected that it would only be set when changed in general usage
regardless.