Determine shader compile flags by config type.
When non-debug, optimizations are applied. When not debug or release with
debug info, names are stripped. The OPTIMIZE parameter to
ds_compile_shaders() now overrides the optimization level rather than
enabling optimizations.

Added ds_config_binary_dir() helper function to get the current binary
directory with the config appended for multi-config generators. This is
used when compiling shaders so the build results with different flags can
co-exist. A workaround in ds_compile_shaders() is used to support different
configs with the ouptut files.