Re-worked explicit clearing of surfaces.
Replaced dsRenderer_clearColorSurface() and
dsRenderer_clearDepthStencilSurface() with dsRenderer_clearAttachments().
Rather than using a transfer operation outside of a render pass, this is
effectively a draw call inside of a render pass. This is simpler for both
OpenGL and Vulkan, and should be either the same speed or faster depending
on the platform.

Removed clear usage flags since it is no longer relevant, and adjusted the
usage flags in Vulkan based on the updated usage assumptions. Added
dsRenderSurfaceUsage flags to allow for more efficient usage.

Currently stubbed out for Metal, and will be implemented in a later commit
with implicit draw calls.