Implemented OBJ model conversion.
Can now use RGBA vertex types in addition to XYZW for flatbuffers.

Primitive type is now provided for each model geometry when converting
models. This allows it to be determined from the file, as well as
supporting multiple primitive types within the same model format. (though
they may not be mixed in an individual piece of geometry)