Cleaned up some module directories.
When multiple sub-libraries are in the same top-level module folder, the
"main" library is now in its own subdirectory. This makes it easier to
distinguish what files are related directly with the library and which are
for a related but separate library.

Simplified some paths. They were originally more explicit due to the old
implementation of the single shared option.