Added dsSceneVectorTextNode.
Holds text and information on how to draw it. Layout is performed by the
item list.

A version number is incremented whenever
dsSceneVectorTextNode_updateLayout() is a called to fully re-layout the
text. Otherwise it's refreshed on draw to ensure that the glyphs are
available.

Fixed copy/paste errors in SceneVectorImageNode.h.