Fixed epsilon checks with clipper.
The epsilon values should respect their values in original floating point
space, so the scale value must be taken into account. This can be different
for X and Y values, so separate the epsilons by axis. The ceiling of the
result is used to guarantee no situation where two points would be
considered epsilon equivalent when converted back to floating point.