Got TestVectorDrawScene running on Android.
Handle optimized rotation properly for shaders in TestVectorDrawScene using
ViewTransform.

Fixes for Android with latest version of Android Studio and gradle.

Always enable shader optimizations on Android. This works around a Qualcomm
driver issue where using smoothstep in the dsComputeTextColor() function
causes the shader to fail to link. When optimizations are enabled it works,
probably due to optimizations inlining the functions. This is in line with
my experience of Qualcomm drivers being the worst I have ever seen on any
platform. Debugging of shaders typically won't be done on mobile platforms,
so this will likely not be a hindrance.