Re-organized JSON for scenes.
SceneResources is now an array of objects with a "type" parameter rather
than arrays for each type.

Removed all "data" elements and instead combine all the elements in a
single object. This is the case for all customizable types, not just for
scene resources.

This has multiple benefits:
* Reduces the amount of horizontal space used.
* Reduces overall complexity of the JSON.
* No longer has a hidden order of processing of SceneResources, which can
  make inter-dependencies difficult to manage.

Documentation will be updated in a followup commit.