Improvements for text signed distance field.
Swapped if checks in the pointInsideGlyph function to optimize for which if
will be hit most often.

Handle ambiguous ordering of X values for *almost* parallel points. Also
check for corner points explicitly, which will handle the case where a top
corner is matched.

Changed some epsilon checks to be able to cache the values slightly to the
left and right of the target point.