Added frustum intersection function to BVH.
Allows for efficient culling from a BVH with a frustum, and more accurate
than attempting to decompose the frustum into an aligned bounding box. This
will also avoid extra tests if a frustum fully encloses the bounds.

Added missing intensity check when visiting the brightest lights. Also
sightly improved some of the logic when triangulating a monotonic polygon
loop.