Added initial light shader code.
Added uniforms for forward lights as well as general lighting functions.
Separate functions are used for directional, point, and spot lights, with
output parameters for the falloff factor and surface direction to the light
to be used to calculate the specular if desired.

Specular has two versions: directional and general. The directional version
uses the half vector, while the general version uses the reflection
direction for both point and spot lights.